added several tests for event form and global form

+++ internal/pkg/models/user.go
@@ -57,8 +57,8 @@ func (ug *DBUserGeneral) GetUserGeneral() UserGeneral {
 
 // Struct describes location point of user
 type LocationPoint struct {
-	Latitude  float64 `json:"lat"`
-	Longitude float64 `json:"lng"`
+	Latitude  float32 `json:"lat"`
+	Longitude float32 `json:"lng"`
 	Accuracy  int     `json:"accuracy, omitempty"`
 }
